Added support in the smart contract and activator to handle multicast groups. The activator is now responsible for assigning the multicast IP to each group
