+++ activator/src/activator.rs
@@ -1,39 +1,29 @@
-use std::collections::HashMap;
-
+use crate::{
+    activator_metrics::ActivatorMetrics,
+    idallocator::IDAllocator,
+    ipblockallocator::IPBlockAllocator,
+    metrics_service::MetricsService,
+    process::{
+        device::process_device_event, exchange::process_exchange_event,
+        location::process_location_event, multicastgroup::process_multicastgroup_event,
+        tunnel::process_tunnel_event, user::process_user_event,
+    },
+    states::devicestate::DeviceState,
+};
 use doublezero_sdk::{
     commands::{
-        device::{
-            activate::ActivateDeviceCommand, deactivate::DeactivateDeviceCommand,
-            get::GetDeviceCommand, list::ListDeviceCommand,
-        },
-        exchange::list::ListExchangeCommand,
-        location::list::ListLocationCommand,
-        multicastgroup::{
-            activate::ActivateMulticastGroupCommand, deactivate::DeactivateMulticastGroupCommand,
-        },
-        tunnel::{
-            activate::ActivateTunnelCommand, deactivate::DeactivateTunnelCommand,
-            list::ListTunnelCommand, reject::RejectTunnelCommand,
-        },
-        user::{
-            activate::ActivateUserCommand, ban::BanUserCommand, deactivate::DeactivateUserCommand,
-            list::ListUserCommand, reject::RejectUserCommand,
-        },
+        device::list::ListDeviceCommand, exchange::list::ListExchangeCommand,
+        location::list::ListLocationCommand, tunnel::list::ListTunnelCommand,
+        user::list::ListUserCommand,
     },
-    ipv4_to_string, networkv4_list_to_string, networkv4_to_string, AccountData, DZClient, Device,
-    DeviceStatus, DoubleZeroClient, Exchange, IpV4, Location, MulticastGroupStatus, Tunnel,
-    TunnelStatus, User, UserStatus, UserType,
+    ipv4_to_string, networkv4_list_to_string, AccountData, DZClient, Device, DeviceStatus,
+    Exchange, Location, TunnelStatus, UserStatus,
 };
 use doublezero_sdk::{GetGlobalConfigCommand, MulticastGroup};
 use solana_sdk::pubkey::Pubkey;
+use std::collections::HashMap;
 use std::thread;
 use std::time::Duration;
-
-use crate::{
-    activator_metrics::ActivatorMetrics, idallocator::IDAllocator,
-    ipblockallocator::IPBlockAllocator, metrics_service::MetricsService,
-    states::devicestate::DeviceState,
-};
 
 pub type DeviceMap = HashMap<Pubkey, DeviceState>;
 
@@ -42,6 +32,7 @@ pub struct Activator {
 
     pub tunnel_tunnel_ids: IDAllocator,
     pub tunnel_tunnel_ips: IPBlockAllocator,
+    pub multicastgroup_tunnel_ips: IPBlockAllocator,
 
     pub user_tunnel_ips: IPBlockAllocator,
     pub devices: DeviceMap,
@@ -87,6 +78,7 @@ impl Activator {
             client,
             tunnel_tunnel_ips: IPBlockAllocator::new(config.tunnel_tunnel_block),
             tunnel_tunnel_ids: IDAllocator::new(0, vec![]),
+            multicastgroup_tunnel_ips: IPBlockAllocator::new(config.multicastgroup_block),
             user_tunnel_ips: IPBlockAllocator::new(config.user_tunnel_block),
             devices: HashMap::new(),
             metrics: ActivatorMetrics::new(metrics_service),
@@ -182,6 +174,7 @@ impl Activator {
         let devices = &mut self.devices;
         let tunnel_tunnel_ips = &mut self.tunnel_tunnel_ips;
         let tunnel_tunnel_ids = &mut self.tunnel_tunnel_ids;
+        let multicastgroup_tunnel_ips = &mut self.multicastgroup_tunnel_ips;
         let user_tunnel_ips = &mut self.user_tunnel_ips;
         let metrics = &self.metrics;
         let locations = &mut self.locations;
@@ -225,8 +218,9 @@ impl Activator {
                         process_multicastgroup_event(
                             client,
                             pubkey,
-                            multicastgroups,
                             multicastgroup,
+                            multicastgroups,
+                            multicastgroup_tunnel_ips,
                             state_transitions,
                         );
                     }
@@ -235,1267 +229,5 @@ impl Activator {
                 metrics.record_metrics(devices, locations, exchanges, state_transitions);
             })?;
         Ok(())
-    }
-}
-
-fn process_device_event(
-    client: &dyn DoubleZeroClient,
-    pubkey: &Pubkey,
-    devices: &mut DeviceMap,
-    device: &Device,
-    state_transitions: &mut HashMap<&'static str, usize>,
-) {
-    match device.status {
-        DeviceStatus::Pending => {
-            print!("New Device {} ", device.code);
-
-            let res = ActivateDeviceCommand {
-                index: device.index,
-            }
-            .execute(client);
-
-            match res {
-                Ok(signature) => {
-                    println!("Activated {}", signature);
-
-                    println!(
-                        "Add Device: {} public_ip: {} dz_prefixes: {} ",
-                        device.code,
-                        ipv4_to_string(&device.public_ip),
-                        networkv4_list_to_string(&device.dz_prefixes)
-                    );
-                    devices.insert(*pubkey, DeviceState::new(device));
-                    *state_transitions
-                        .entry("device-pending-to-activated")
-                        .or_insert(0) += 1;
-                }
-                Err(e) => println!("Error: {}", e),
-            }
-        }
-        DeviceStatus::Activated => {
-            if !devices.contains_key(pubkey) {
-                println!(
-                    "Add Device: {} public_ip: {} dz_prefixes: {} ",
-                    device.code,
-                    ipv4_to_string(&device.public_ip),
-                    networkv4_list_to_string(&device.dz_prefixes)
-                );
-
-                devices.insert(*pubkey, DeviceState::new(device));
-            } else {
-                let device_state = devices.get_mut(pubkey).unwrap();
-                device_state.update(device);
-            }
-        }
-        DeviceStatus::Deleting => {
-            print!("Deleting Device {} ", device.code);
-
-            let res = DeactivateDeviceCommand {
-                index: device.index,
-                owner: device.owner,
-            }
-            .execute(client);
-
-            match res {
-                Ok(signature) => {
-                    println!("Deactivated {}", signature);
-                    devices.remove(pubkey);
-                    *state_transitions
-                        .entry("device-deleting-to-deactivated")
-                        .or_insert(0) += 1;
-                }
-                Err(e) => println!("Error: {}", e),
-            }
-        }
-        _ => {}
-    }
-}
-
-fn process_tunnel_event(
-    client: &dyn DoubleZeroClient,
-    tunnel_tunnel_ips: &mut IPBlockAllocator,
-    tunnel_tunnel_ids: &mut IDAllocator,
-    tunnel: &Tunnel,
-    state_transitions: &mut HashMap<&'static str, usize>,
-) {
-    match tunnel.status {
-        TunnelStatus::Pending => {
-            print!("New Tunnel {} ", tunnel.code);
-
-            match tunnel_tunnel_ips.next_available_block(0, 2) {
-                Some(tunnel_net) => {
-                    let tunnel_id = tunnel_tunnel_ids.next_available();
-
-                    let res = ActivateTunnelCommand {
-                        index: tunnel.index,
-                        tunnel_id,
-                        tunnel_net,
-                    }
-                    .execute(client);
-
-                    match res {
-                        Ok(signature) => {
-                            println!("Activated {}", signature);
-                            *state_transitions
-                                .entry("tunnel-pending-to-activated")
-                                .or_insert(0) += 1;
-                        }
-                        Err(e) => println!("Error: activate_tunnel: {}", e),
-                    }
-                }
-                None => {
-                    println!("Error: No available tunnel block");
-
-                    let res = RejectTunnelCommand {
-                        index: tunnel.index,
-                        reason: "Error: No available tunnel block".to_string(),
-                    }
-                    .execute(client);
-
-                    match res {
-                        Ok(signature) => {
-                            println!("Rejected {}", signature);
-                            *state_transitions
-                                .entry("tunnel-pending-to-rejected")
-                                .or_insert(0) += 1;
-                        }
-                        Err(e) => println!("Error: reject_tunnel: {}", e),
-                    }
-                }
-            }
-        }
-        TunnelStatus::Deleting => {
-            print!("Deleting Tunnel {} ", tunnel.code);
-
-            tunnel_tunnel_ids.unassign(tunnel.tunnel_id);
-            tunnel_tunnel_ips.unassign_block(tunnel.tunnel_net);
-
-            let res = DeactivateTunnelCommand {
-                index: tunnel.index,
-                owner: tunnel.owner,
-            }
-            .execute(client);
-
-            match res {
-                Ok(signature) => {
-                    println!("Deactivated {}", signature);
-                    *state_transitions
-                        .entry("tunnel-deleting-to-deactivated")
-                        .or_insert(0) += 1;
-                }
-                Err(e) => println!("Error: deactivate_tunnel: {}", e),
-            }
-        }
-        _ => {}
-    }
-}
-
-fn process_user_event(
-    client: &dyn DoubleZeroClient,
-    pubkey: &Pubkey,
-    devices: &mut DeviceMap,
-    user_tunnel_ips: &mut IPBlockAllocator,
-    tunnel_tunnel_ids: &mut IDAllocator,
-    user: &User,
-    state_transitions: &mut HashMap<&'static str, usize>,
-) {
-    match user.status {
-        // Create User
-        UserStatus::Pending => {
-            print!("Activating User   {} ", ipv4_to_string(&user.client_ip));
-            // Load Device if not exists
-            if !devices.contains_key(&user.device_pk) {
-                let res = GetDeviceCommand {
-                    pubkey_or_code: user.device_pk.to_string(),
-                }
-                .execute(client);
-
-                match res {
-                    Ok((_, device)) => {
-                        println!(
-                            "Add Device: {} public_ip: {} dz_prefixes: {} ",
-                            device.code,
-                            ipv4_to_string(&device.public_ip),
-                            networkv4_list_to_string(&device.dz_prefixes)
-                        );
-
-                        devices.insert(*pubkey, DeviceState::new(&device));
-                    }
-                    Err(e) => {
-                        println!("Error: {}", e);
-                    }
-                }
-            }
-
-            match devices.get_mut(&user.device_pk) {
-                Some(device_state) => {
-                    print!("for {} ", device_state.device.code);
-
-                    match user_tunnel_ips.next_available_block(0, 2) {
-                        Some(tunnel_net) => {
-                            print!("tunnel_net: {} ", networkv4_to_string(&tunnel_net));
-
-                            let mut tunnel_id: u16 = 0;
-                            let mut dz_ip: IpV4 = [0, 0, 0, 0];
-
-                            match user.user_type {
-                                UserType::IBRLWithAllocatedIP | UserType::EdgeFiltering => {
-                                    if let Some((xtunnel_id, xdz_ip)) = device_state.get_next() {
-                                        tunnel_id = xtunnel_id;
-                                        dz_ip = xdz_ip;
-                                    }
-                                }
-                                UserType::IBRL => {
-                                    tunnel_id = device_state.get_next_tunnel_id().unwrap();
-                                    dz_ip = user.client_ip;
-                                }
-                                UserType::Multicast => {}
-                            }
-
-                            if tunnel_id == 0 {
-                                eprintln!("Error: No available tunnel block");
-
-                                let res = RejectUserCommand {
-                                    index: user.index,
-                                    reason: "Error: No available tunnel block".to_string(),
-                                }
-                                .execute(client);
-
-                                match res {
-                                    Ok(signature) => {
-                                        println!("Rejected {}", signature);
-                                        *state_transitions
-                                            .entry("user-pending-to-rejected")
-                                            .or_insert(0) += 1;
-                                    }
-                                    Err(e) => println!("Error: {}", e),
-                                }
-                                return;
-                            }
-
-                            print!(
-                                "tunnel_id: {} dz_ip: {} ",
-                                tunnel_id,
-                                ipv4_to_string(&dz_ip)
-                            );
-
-                            let res = ActivateUserCommand {
-                                index: user.index,
-                                tunnel_id,
-                                tunnel_net,
-                                dz_ip,
-                            }
-                            .execute(client);
-
-                            match res {
-                                Ok(signature) => {
-                                    println!("Activated   {}", signature);
-                                    *state_transitions
-                                        .entry("user-pending-to-activated")
-                                        .or_insert(0) += 1;
-                                }
-                                Err(e) => println!("Error: {}", e),
-                            }
-                        }
-                        None => {
-                            println!("Error: No available user block");
-
-                            let res = RejectUserCommand {
-                                index: user.index,
-                                reason: "Error: No available user block".to_string(),
-                            }
-                            .execute(client);
-
-                            match res {
-                                Ok(signature) => {
-                                    println!("Rejected {}", signature);
-                                    *state_transitions
-                                        .entry("user-pending-to-rejected")
-                                        .or_insert(0) += 1;
-                                }
-                                Err(e) => println!("Error: {}", e),
-                            }
-                        }
-                    }
-                }
-                None => {
-                    eprintln!("Error: Device not found {}", user.device_pk);
-
-                    let res = RejectUserCommand {
-                        index: user.index,
-                        reason: "Error: Device not found".to_string(),
-                    }
-                    .execute(client);
-
-                    match res {
-                        Ok(signature) => {
-                            println!("Rejected {}", signature);
-                            *state_transitions
-                                .entry("user-pending-to-rejected")
-                                .or_insert(0) += 1;
-                        }
-                        Err(e) => println!("Error: {}", e),
-                    }
-                }
-            }
-        }
-        // Delete User
-        UserStatus::Deleting | UserStatus::PendingBan => {
-            print!("Deactivating User {} ", ipv4_to_string(&user.client_ip));
-
-            if let Some(device_state) = devices.get_mut(&user.device_pk) {
-                print!("for {} ", device_state.device.code);
-
-                print!(
-                    "tunnel_net: {} tunnel_id: {} dz_ip: {} ",
-                    networkv4_to_string(&user.tunnel_net),
-                    user.tunnel_id,
-                    ipv4_to_string(&user.dz_ip)
-                );
-
-                if user.tunnel_id != 0 {
-                    tunnel_tunnel_ids.unassign(user.tunnel_id);
-                }
-                if user.tunnel_net != ([0, 0, 0, 0], 0) {
-                    user_tunnel_ips.unassign_block(user.tunnel_net);
-                }
-                if user.dz_ip != [0, 0, 0, 0] {
-                    device_state.release(user.dz_ip, user.tunnel_id);
-                }
-
-                if user.status == UserStatus::Deleting {
-                    let res = DeactivateUserCommand {
-                        index: user.index,
-                        owner: user.owner,
-                    }
-                    .execute(client);
-
-                    match res {
-                        Ok(signature) => {
-                            println!("Deactivated {}", signature);
-                            *state_transitions
-                                .entry("user-deleting-to-deactivated")
-                                .or_insert(0) += 1;
-                        }
-                        Err(e) => println!("Error: {}", e),
-                    }
-                } else if user.status == UserStatus::PendingBan {
-                    let res = BanUserCommand { index: user.index }.execute(client);
-
-                    match res {
-                        Ok(signature) => {
-                            println!("Banned {}", signature);
-                            *state_transitions
-                                .entry("user-pending-ban-to-banned")
-                                .or_insert(0) += 1;
-                        }
-                        Err(e) => println!("Error: {}", e),
-                    }
-                }
-            }
-        }
-        _ => {}
-    }
-}
-
-fn process_location_event(
-    pubkey: &Pubkey,
-    locations: &mut HashMap<Pubkey, Location>,
-    location: &Location,
-) {
-    locations.insert(*pubkey, location.clone());
-}
-
-fn process_exchange_event(
-    pubkey: &Pubkey,
-    exchanges: &mut HashMap<Pubkey, Exchange>,
-    exchange: &Exchange,
-) {
-    exchanges.insert(*pubkey, exchange.clone());
-}
-
-fn process_multicastgroup_event(
-    client: &dyn DoubleZeroClient,
-    pubkey: &Pubkey,
-    multicastgroups: &mut HashMap<Pubkey, MulticastGroup>,
-    multicastgroup: &MulticastGroup,
-    state_transitions: &mut HashMap<&'static str, usize>,
-) {
-    match multicastgroup.status {
-        MulticastGroupStatus::Pending => {
-            print!("New MulticastGroup {} ", multicastgroup.code);
-
-            let res = ActivateMulticastGroupCommand {
-                index: multicastgroup.index,
-            }
-            .execute(client);
-
-            match res {
-                Ok(signature) => {
-                    println!("Activated {}", signature);
-
-                    println!("Add MulticastGroup: {} ", multicastgroup.code,);
-                    multicastgroups.insert(*pubkey, multicastgroup.clone());
-                    *state_transitions
-                        .entry("multicastgroup-pending-to-activated")
-                        .or_insert(0) += 1;
-                }
-                Err(e) => println!("Error: {}", e),
-            }
-        }
-        MulticastGroupStatus::Activated => {
-            if !multicastgroups.contains_key(pubkey) {
-                println!("Add MulticastGroup: {} ", multicastgroup.code,);
-
-                multicastgroups.insert(*pubkey, multicastgroup.clone());
-            }
-        }
-        MulticastGroupStatus::Deleting => {
-            print!("Deleting MulticastGroup {} ", multicastgroup.code);
-
-            let res = DeactivateMulticastGroupCommand {
-                index: multicastgroup.index,
-                owner: multicastgroup.owner,
-            }
-            .execute(client);
-
-            match res {
-                Ok(signature) => {
-                    println!("Deactivated {}", signature);
-                    multicastgroups.remove(pubkey);
-                    *state_transitions
-                        .entry("multicastgroup-deleting-to-deactivated")
-                        .or_insert(0) += 1;
-                }
-                Err(e) => println!("Error: {}", e),
-            }
-        }
-        _ => {}
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use doublezero_sdk::{
-        AccountType, DeviceType, ExchangeStatus, LocationStatus, MockDoubleZeroClient,
-        TunnelTunnelType, UserCYOA,
-    };
-    use doublezero_sla_program::{
-        instructions::DoubleZeroInstruction,
-        pda::get_device_pda,
-        pda::get_globalstate_pda,
-        pda::get_tunnel_pda,
-        pda::get_user_pda,
-        processors::{
-            device::{activate::DeviceActivateArgs, deactivate::DeviceDeactivateArgs},
-            tunnel::{
-                activate::TunnelActivateArgs, deactivate::TunnelDeactivateArgs,
-                reject::TunnelRejectArgs,
-            },
-            user::{
-                activate::UserActivateArgs, ban::UserBanArgs, deactivate::UserDeactivateArgs,
-                reject::UserRejectArgs,
-            },
-        },
-        state::globalstate::GlobalState,
-    };
-    use mockall::{predicate, Sequence};
-    use solana_sdk::signature::Signature;
-
-    use super::*;
-
-    fn create_test_client() -> MockDoubleZeroClient {
-        let mut client = MockDoubleZeroClient::new();
-
-        // Program ID
-        let program_id = Pubkey::new_unique();
-        client.expect_get_program_id().returning(move || program_id);
-
-        // Global State
-        let (globalstate_pubkey, _) = get_globalstate_pda(&program_id);
-        let globalstate = GlobalState {
-            account_type: AccountType::GlobalState,
-            bump_seed: 0,
-            account_index: 0,
-            foundation_allowlist: vec![],
-            device_allowlist: vec![],
-            user_allowlist: vec![],
-        };
-
-        client
-            .expect_get()
-            .with(predicate::eq(globalstate_pubkey))
-            .returning(move |_| Ok(AccountData::GlobalState(globalstate.clone())));
-
-        client
-    }
-
-    fn get_device_bump_seed(client: &MockDoubleZeroClient) -> u8 {
-        let (_, bump_seed) = get_device_pda(&client.get_program_id(), 0);
-        bump_seed
-    }
-
-    fn get_tunnel_bump_seed(client: &MockDoubleZeroClient) -> u8 {
-        let (_, bump_seed) = get_tunnel_pda(&client.get_program_id(), 0);
-        bump_seed
-    }
-
-    fn get_user_bump_seed(client: &MockDoubleZeroClient) -> u8 {
-        let (_, bump_seed) = get_user_pda(&client.get_program_id(), 0);
-        bump_seed
-    }
-
-    #[test]
-    fn test_process_device_event_pending_to_deleted() {
-        let mut seq = Sequence::new();
-        let mut devices = HashMap::new();
-        let mut client = create_test_client();
-
-        let device_pubkey = Pubkey::new_unique();
-        let mut device = Device {
-            account_type: AccountType::Device,
-            owner: Pubkey::new_unique(),
-            index: 0,
-            bump_seed: get_device_bump_seed(&client),
-            location_pk: Pubkey::new_unique(),
-            exchange_pk: Pubkey::new_unique(),
-            device_type: DeviceType::Switch,
-            public_ip: [192, 168, 1, 1],
-            status: DeviceStatus::Pending,
-            code: "TestDevice".to_string(),
-            dz_prefixes: vec![([10, 0, 0, 1], 24), ([10, 0, 1, 1], 24)],
-        };
-
-        let mut state_transitions: HashMap<&'static str, usize> = HashMap::new();
-
-        client
-            .expect_execute_transaction()
-            .times(1)
-            .in_sequence(&mut seq)
-            .with(
-                predicate::eq(DoubleZeroInstruction::ActivateDevice(DeviceActivateArgs {
-                    index: device.index,
-                    bump_seed: device.bump_seed,
-                })),
-                predicate::always(),
-            )
-            .returning(|_, _| Ok(Signature::new_unique()));
-
-        process_device_event(
-            &client,
-            &device_pubkey,
-            &mut devices,
-            &device,
-            &mut state_transitions,
-        );
-
-        assert!(devices.contains_key(&device_pubkey));
-        assert_eq!(devices.get(&device_pubkey).unwrap().device, device);
-
-        device.status = DeviceStatus::Deleting;
-
-        let mut client = create_test_client();
-        let device_bump_seed = get_device_bump_seed(&client);
-        client
-            .expect_execute_transaction()
-            .times(1)
-            .in_sequence(&mut seq)
-            .with(
-                predicate::eq(DoubleZeroInstruction::DeactivateDevice(
-                    DeviceDeactivateArgs {
-                        index: device.index,
-                        bump_seed: device_bump_seed,
-                    },
-                )),
-                predicate::always(),
-            )
-            .returning(|_, _| Ok(Signature::new_unique()));
-
-        process_device_event(
-            &client,
-            &device_pubkey,
-            &mut devices,
-            &device,
-            &mut state_transitions,
-        );
-        assert!(!devices.contains_key(&device_pubkey));
-        assert_eq!(state_transitions.len(), 2);
-        assert_eq!(state_transitions["device-pending-to-activated"], 1);
-        assert_eq!(state_transitions["device-deleting-to-deactivated"], 1);
-    }
-
-    #[test]
-    fn test_process_device_event_activated() {
-        let mut devices = HashMap::new();
-        let client = create_test_client();
-        let pubkey = Pubkey::new_unique();
-
-        let mut device = Device {
-            account_type: AccountType::Device,
-            owner: Pubkey::new_unique(),
-            index: 0,
-            bump_seed: get_device_bump_seed(&client),
-            location_pk: Pubkey::new_unique(),
-            exchange_pk: Pubkey::new_unique(),
-            device_type: DeviceType::Switch,
-            public_ip: [192, 168, 1, 1],
-            status: DeviceStatus::Activated,
-            code: "TestDevice".to_string(),
-            dz_prefixes: vec![([10, 0, 0, 1], 24)],
-        };
-
-        let mut state_transitions: HashMap<&'static str, usize> = HashMap::new();
-
-        process_device_event(
-            &client,
-            &pubkey,
-            &mut devices,
-            &device,
-            &mut state_transitions,
-        );
-
-        assert!(devices.contains_key(&pubkey));
-        assert_eq!(devices.get(&pubkey).unwrap().device, device);
-
-        device.dz_prefixes.push(([10, 0, 1, 1], 24));
-        process_device_event(
-            &client,
-            &pubkey,
-            &mut devices,
-            &device,
-            &mut state_transitions,
-        );
-
-        assert!(devices.contains_key(&pubkey));
-        assert_eq!(devices.get(&pubkey).unwrap().device, device);
-
-        assert_eq!(state_transitions.len(), 0);
-    }
-
-    #[test]
-    fn test_process_tunnel_event_pending_to_deleted() {
-        let mut seq = Sequence::new();
-        let mut tunnel_tunnel_ips = IPBlockAllocator::new(([10, 0, 0, 0], 16));
-        let mut tunnel_tunnel_ids = IDAllocator::new(500, vec![500, 501, 503]);
-        let mut client = create_test_client();
-
-        let tunnel = Tunnel {
-            account_type: AccountType::Tunnel,
-            owner: Pubkey::new_unique(),
-            index: 0,
-            bump_seed: get_tunnel_bump_seed(&client),
-            side_a_pk: Pubkey::new_unique(),
-            side_z_pk: Pubkey::new_unique(),
-            tunnel_type: TunnelTunnelType::MPLSoGRE,
-            bandwidth: 10_000_000_000,
-            mtu: 1500,
-            delay_ns: 100,
-            jitter_ns: 100,
-            tunnel_id: 1,
-            tunnel_net: ([0, 0, 0, 0], 0),
-            status: TunnelStatus::Pending,
-            code: "TestTunnel".to_string(),
-        };
-
-        client
-            .expect_execute_transaction()
-            .times(1)
-            .in_sequence(&mut seq)
-            .with(
-                predicate::eq(DoubleZeroInstruction::ActivateTunnel(TunnelActivateArgs {
-                    index: tunnel.index,
-                    bump_seed: tunnel.bump_seed,
-                    tunnel_id: 502,
-                    tunnel_net: ([10, 0, 0, 0], 31),
-                })),
-                predicate::always(),
-            )
-            .returning(|_, _| Ok(Signature::new_unique()));
-
-        let mut state_transitions: HashMap<&'static str, usize> = HashMap::new();
-
-        process_tunnel_event(
-            &client,
-            &mut tunnel_tunnel_ips,
-            &mut tunnel_tunnel_ids,
-            &tunnel,
-            &mut state_transitions,
-        );
-
-        assert!(tunnel_tunnel_ids.assigned.contains(&502_u16));
-        assert!(tunnel_tunnel_ips.contains([10, 0, 0, 42]));
-
-        let mut tunnel = tunnel.clone();
-        tunnel.status = TunnelStatus::Deleting;
-        tunnel.tunnel_id = 502;
-        tunnel.tunnel_net = ([10, 0, 0, 0], 31);
-
-        client
-            .expect_execute_transaction()
-            .times(1)
-            .in_sequence(&mut seq)
-            .with(
-                predicate::eq(DoubleZeroInstruction::DeactivateTunnel(
-                    TunnelDeactivateArgs {
-                        index: tunnel.index,
-                        bump_seed: tunnel.bump_seed,
-                    },
-                )),
-                predicate::always(),
-            )
-            .returning(|_, _| Ok(Signature::new_unique()));
-
-        let assigned_ips = tunnel_tunnel_ips.assigned_ips.clone();
-
-        process_tunnel_event(
-            &client,
-            &mut tunnel_tunnel_ips,
-            &mut tunnel_tunnel_ids,
-            &tunnel,
-            &mut state_transitions,
-        );
-
-        assert!(!tunnel_tunnel_ids.assigned.contains(&502_u16));
-        assert_ne!(tunnel_tunnel_ips.assigned_ips, assigned_ips);
-
-        assert_eq!(state_transitions.len(), 2);
-        assert_eq!(state_transitions["tunnel-pending-to-activated"], 1);
-        assert_eq!(state_transitions["tunnel-deleting-to-deactivated"], 1);
-    }
-
-    #[test]
-    fn test_process_tunnel_event_rejected() {
-        let mut seq = Sequence::new();
-        let mut tunnel_tunnel_ips = IPBlockAllocator::new(([10, 0, 0, 0], 32));
-        let mut tunnel_tunnel_ids = IDAllocator::new(500, vec![500, 501, 503]);
-        let mut client = create_test_client();
-
-        let tunnel = Tunnel {
-            account_type: AccountType::Tunnel,
-            owner: Pubkey::new_unique(),
-            index: 0,
-            bump_seed: get_tunnel_bump_seed(&client),
-            side_a_pk: Pubkey::new_unique(),
-            side_z_pk: Pubkey::new_unique(),
-            tunnel_type: TunnelTunnelType::MPLSoGRE,
-            bandwidth: 10_000_000_000,
-            mtu: 1500,
-            delay_ns: 100,
-            jitter_ns: 100,
-            tunnel_id: 1,
-            tunnel_net: ([0, 0, 0, 0], 0),
-            status: TunnelStatus::Pending,
-            code: "TestTunnel".to_string(),
-        };
-
-        let _ = tunnel_tunnel_ips.next_available_block(0, 2);
-
-        client
-            .expect_execute_transaction()
-            .times(1)
-            .in_sequence(&mut seq)
-            .with(
-                predicate::eq(DoubleZeroInstruction::RejectTunnel(TunnelRejectArgs {
-                    index: tunnel.index,
-                    bump_seed: tunnel.bump_seed,
-                    reason: "Error: No available tunnel block".to_string(),
-                })),
-                predicate::always(),
-            )
-            .returning(|_, _| Ok(Signature::new_unique()));
-
-        let mut state_transitions: HashMap<&'static str, usize> = HashMap::new();
-
-        process_tunnel_event(
-            &client,
-            &mut tunnel_tunnel_ips,
-            &mut tunnel_tunnel_ids,
-            &tunnel,
-            &mut state_transitions,
-        );
-
-        assert_eq!(state_transitions.len(), 1);
-        assert_eq!(state_transitions["tunnel-pending-to-rejected"], 1);
-    }
-
-    fn do_test_process_user_event_pending_to_activated(
-        user_type: UserType,
-        expected_dz_ip: Option<IpV4>,
-    ) {
-        let mut seq = Sequence::new();
-        let mut user_tunnel_ips = IPBlockAllocator::new(([10, 0, 0, 0], 16));
-        let mut tunnel_tunnel_ids = IDAllocator::new(100, vec![100, 101, 102]);
-        let mut client = create_test_client();
-
-        let device_pubkey = Pubkey::new_unique();
-        let device = Device {
-            account_type: AccountType::Device,
-            owner: Pubkey::new_unique(),
-            index: 0,
-            bump_seed: get_device_bump_seed(&client),
-            location_pk: Pubkey::new_unique(),
-            exchange_pk: Pubkey::new_unique(),
-            device_type: DeviceType::Switch,
-            public_ip: [192, 168, 1, 2],
-            status: DeviceStatus::Activated,
-            code: "TestDevice".to_string(),
-            dz_prefixes: vec![([10, 0, 0, 1], 24)],
-        };
-
-        let user_pubkey = Pubkey::new_unique();
-        let user = User {
-            account_type: AccountType::User,
-            owner: Pubkey::new_unique(),
-            index: 0,
-            bump_seed: get_user_bump_seed(&client),
-            user_type,
-            tenant_pk: Pubkey::new_unique(),
-            device_pk: device_pubkey,
-            cyoa_type: UserCYOA::GREOverDIA,
-            client_ip: [192, 168, 1, 1],
-            dz_ip: [0, 0, 0, 0],
-            tunnel_id: 0,
-            tunnel_net: ([0, 0, 0, 0], 0),
-            status: UserStatus::Pending,
-        };
-
-        client
-            .expect_execute_transaction()
-            .times(1)
-            .in_sequence(&mut seq)
-            .with(
-                predicate::eq(DoubleZeroInstruction::ActivateUser(UserActivateArgs {
-                    index: user.index,
-                    bump_seed: user.bump_seed,
-                    tunnel_id: 500,
-                    tunnel_net: ([10, 0, 0, 0], 31),
-                    dz_ip: expected_dz_ip.unwrap_or([0, 0, 0, 0]),
-                })),
-                predicate::always(),
-            )
-            .returning(|_, _| Ok(Signature::new_unique()));
-
-        let mut state_transitions: HashMap<&'static str, usize> = HashMap::new();
-
-        let mut devices = HashMap::new();
-        devices.insert(device_pubkey, DeviceState::new(&device));
-
-        process_user_event(
-            &client,
-            &user_pubkey,
-            &mut devices,
-            &mut user_tunnel_ips,
-            &mut tunnel_tunnel_ids,
-            &user,
-            &mut state_transitions,
-        );
-
-        assert!(user_tunnel_ips.assigned_ips.len() > 0);
-        assert!(tunnel_tunnel_ids.assigned.len() > 0);
-
-        assert_eq!(state_transitions.len(), 1);
-        assert_eq!(state_transitions["user-pending-to-activated"], 1);
-    }
-
-    #[test]
-    fn test_process_user_event_pending_to_activated_ibrl() {
-        do_test_process_user_event_pending_to_activated(UserType::IBRL, Some([192, 168, 1, 1]));
-    }
-
-    #[test]
-    fn test_process_user_event_pending_to_activated_ibrl_with_allocated_ip() {
-        do_test_process_user_event_pending_to_activated(
-            UserType::IBRLWithAllocatedIP,
-            Some([10, 0, 0, 1]),
-        );
-    }
-
-    #[test]
-    fn test_process_user_event_pending_to_activated_edge_filtering() {
-        do_test_process_user_event_pending_to_activated(
-            UserType::EdgeFiltering,
-            Some([10, 0, 0, 1]),
-        );
-    }
-
-    #[test]
-    fn test_process_user_event_pending_to_rejected_by_get_device() {
-        let mut seq = Sequence::new();
-        let mut user_tunnel_ips = IPBlockAllocator::new(([10, 0, 0, 0], 32));
-        let mut tunnel_tunnel_ids = IDAllocator::new(100, vec![100, 101, 102]);
-        let mut client = create_test_client();
-
-        let device_pubkey = Pubkey::new_unique();
-
-        let user_pubkey = Pubkey::new_unique();
-        let user = User {
-            account_type: AccountType::User,
-            owner: Pubkey::new_unique(),
-            index: 0,
-            bump_seed: get_user_bump_seed(&client),
-            user_type: UserType::IBRLWithAllocatedIP,
-            tenant_pk: Pubkey::new_unique(),
-            device_pk: device_pubkey,
-            cyoa_type: UserCYOA::GREOverDIA,
-            client_ip: [192, 168, 1, 1],
-            dz_ip: [0, 0, 0, 0],
-            tunnel_id: 0,
-            tunnel_net: ([0, 0, 0, 0], 0),
-            status: UserStatus::Pending,
-        };
-
-        client
-            .expect_get()
-            .times(1)
-            .in_sequence(&mut seq)
-            .with(predicate::eq(device_pubkey))
-            .returning(|_| Err(eyre::eyre!("Device not found")));
-
-        client
-            .expect_execute_transaction()
-            .times(1)
-            .in_sequence(&mut seq)
-            .with(
-                predicate::eq(DoubleZeroInstruction::RejectUser(UserRejectArgs {
-                    index: user.index,
-                    bump_seed: user.bump_seed,
-                    reason: "Error: Device not found".to_string(),
-                })),
-                predicate::always(),
-            )
-            .returning(|_, _| Ok(Signature::new_unique()));
-
-        let mut state_transitions: HashMap<&'static str, usize> = HashMap::new();
-
-        let mut devices = HashMap::new();
-
-        process_user_event(
-            &client,
-            &user_pubkey,
-            &mut devices,
-            &mut user_tunnel_ips,
-            &mut tunnel_tunnel_ids,
-            &user,
-            &mut state_transitions,
-        );
-
-        assert_eq!(state_transitions.len(), 1);
-        assert_eq!(state_transitions["user-pending-to-rejected"], 1);
-    }
-
-    #[test]
-    fn test_process_user_event_pending_to_rejected_by_no_tunnel_block() {
-        let mut seq = Sequence::new();
-        let mut user_tunnel_ips = IPBlockAllocator::new(([10, 0, 0, 0], 32));
-        let mut tunnel_tunnel_ids = IDAllocator::new(100, vec![100, 101, 102]);
-        let mut client = create_test_client();
-
-        let device_pubkey = Pubkey::new_unique();
-        let device = Device {
-            account_type: AccountType::Device,
-            owner: Pubkey::new_unique(),
-            index: 0,
-            bump_seed: get_device_bump_seed(&client),
-            location_pk: Pubkey::new_unique(),
-            exchange_pk: Pubkey::new_unique(),
-            device_type: DeviceType::Switch,
-            public_ip: [192, 168, 1, 2],
-            status: DeviceStatus::Activated,
-            code: "TestDevice".to_string(),
-            dz_prefixes: vec![([10, 0, 0, 0], 32)],
-        };
-
-        let user_pubkey = Pubkey::new_unique();
-        let user = User {
-            account_type: AccountType::User,
-            owner: Pubkey::new_unique(),
-            index: 0,
-            bump_seed: get_user_bump_seed(&client),
-            user_type: UserType::IBRLWithAllocatedIP,
-            tenant_pk: Pubkey::new_unique(),
-            device_pk: device_pubkey,
-            cyoa_type: UserCYOA::GREOverDIA,
-            client_ip: [192, 168, 1, 1],
-            dz_ip: [0, 0, 0, 0],
-            tunnel_id: 0,
-            tunnel_net: ([0, 0, 0, 0], 0),
-            status: UserStatus::Pending,
-        };
-
-        client
-            .expect_execute_transaction()
-            .times(1)
-            .in_sequence(&mut seq)
-            .with(
-                predicate::eq(DoubleZeroInstruction::RejectUser(UserRejectArgs {
-                    index: user.index,
-                    bump_seed: user.bump_seed,
-                    reason: "Error: No available tunnel block".to_string(),
-                })),
-                predicate::always(),
-            )
-            .returning(|_, _| Ok(Signature::new_unique()));
-
-        let mut state_transitions: HashMap<&'static str, usize> = HashMap::new();
-
-        let mut devices = HashMap::new();
-        let device2 = device.clone();
-        devices.insert(device_pubkey, DeviceState::new(&device2));
-
-        // allocate the only ip
-        assert_ne!(
-            devices.get_mut(&device_pubkey).unwrap().dz_ips[0].next_available_block(1, 1),
-            None
-        );
-
-        process_user_event(
-            &client,
-            &user_pubkey,
-            &mut devices,
-            &mut user_tunnel_ips,
-            &mut tunnel_tunnel_ids,
-            &user,
-            &mut state_transitions,
-        );
-
-        assert_eq!(state_transitions.len(), 1);
-        assert_eq!(state_transitions["user-pending-to-rejected"], 1);
-    }
-
-    #[test]
-    fn test_process_user_event_pending_to_rejected_by_no_user_block() {
-        let mut seq = Sequence::new();
-        let mut user_tunnel_ips = IPBlockAllocator::new(([10, 0, 0, 0], 32));
-        let mut tunnel_tunnel_ids = IDAllocator::new(100, vec![100, 101, 102]);
-        let mut client = create_test_client();
-
-        // eat a blocok
-        let _ = user_tunnel_ips.next_available_block(0, 2);
-
-        let device_pubkey = Pubkey::new_unique();
-        let device = Device {
-            account_type: AccountType::Device,
-            owner: Pubkey::new_unique(),
-            index: 0,
-            bump_seed: get_device_bump_seed(&client),
-            location_pk: Pubkey::new_unique(),
-            exchange_pk: Pubkey::new_unique(),
-            device_type: DeviceType::Switch,
-            public_ip: [192, 168, 1, 2],
-            status: DeviceStatus::Activated,
-            code: "TestDevice".to_string(),
-            dz_prefixes: vec![([10, 0, 0, 1], 24)],
-        };
-
-        let pubkey = Pubkey::new_unique();
-        let user = User {
-            account_type: AccountType::User,
-            owner: Pubkey::new_unique(),
-            index: 0,
-            bump_seed: get_user_bump_seed(&client),
-            user_type: UserType::IBRLWithAllocatedIP,
-            tenant_pk: Pubkey::new_unique(),
-            device_pk: device_pubkey,
-            cyoa_type: UserCYOA::GREOverDIA,
-            client_ip: [192, 168, 1, 1],
-            dz_ip: [0, 0, 0, 0],
-            tunnel_id: 0,
-            tunnel_net: ([0, 0, 0, 0], 0),
-            status: UserStatus::Pending,
-        };
-
-        client
-            .expect_execute_transaction()
-            .times(1)
-            .in_sequence(&mut seq)
-            .with(
-                predicate::eq(DoubleZeroInstruction::RejectUser(UserRejectArgs {
-                    index: user.index,
-                    bump_seed: user.bump_seed,
-                    reason: "Error: No available user block".to_string(),
-                })),
-                predicate::always(),
-            )
-            .returning(|_, _| Ok(Signature::new_unique()));
-
-        let mut state_transitions: HashMap<&'static str, usize> = HashMap::new();
-
-        let mut devices = HashMap::new();
-        let device2 = device.clone();
-        devices.insert(device_pubkey, DeviceState::new(&device2));
-
-        process_user_event(
-            &client,
-            &pubkey,
-            &mut devices,
-            &mut user_tunnel_ips,
-            &mut tunnel_tunnel_ids,
-            &user,
-            &mut state_transitions,
-        );
-
-        assert_eq!(state_transitions.len(), 1);
-        assert_eq!(state_transitions["user-pending-to-rejected"], 1);
-    }
-
-    fn do_test_process_user_event_deleting_or_pending_ban<F>(
-        user_status: UserStatus,
-        func: F,
-        state_transition: &'static str,
-    ) where
-        F: Fn(&mut MockDoubleZeroClient, &User, &mut Sequence) -> (),
-    {
-        assert!(user_status == UserStatus::Deleting || user_status == UserStatus::PendingBan);
-
-        let mut seq = Sequence::new();
-        let mut devices = HashMap::new();
-        let mut user_tunnel_ips = IPBlockAllocator::new(([10, 0, 0, 0], 16));
-        let mut tunnel_tunnel_ids = IDAllocator::new(100, vec![100, 101, 102]);
-        let mut client = create_test_client();
-
-        let mut state_transitions: HashMap<&'static str, usize> = HashMap::new();
-
-        let pubkey = Pubkey::new_unique();
-        let user = User {
-            account_type: AccountType::User,
-            owner: Pubkey::new_unique(),
-            index: 0,
-            bump_seed: get_user_bump_seed(&client),
-            user_type: UserType::IBRLWithAllocatedIP,
-            tenant_pk: Pubkey::new_unique(),
-            device_pk: pubkey,
-            cyoa_type: UserCYOA::GREOverDIA,
-            client_ip: [192, 168, 1, 1],
-            dz_ip: [0, 0, 0, 0],
-            tunnel_id: 102,
-            tunnel_net: ([10, 0, 0, 0], 31),
-            status: user_status,
-        };
-
-        let device = Device {
-            account_type: AccountType::Device,
-            owner: Pubkey::new_unique(),
-            index: 0,
-            bump_seed: get_device_bump_seed(&client),
-            location_pk: Pubkey::new_unique(),
-            exchange_pk: Pubkey::new_unique(),
-            device_type: DeviceType::Switch,
-            public_ip: [192, 168, 1, 2],
-            status: DeviceStatus::Activated,
-            code: "TestDevice".to_string(),
-            dz_prefixes: vec![([11, 0, 0, 0], 16)],
-        };
-
-        devices.insert(pubkey, DeviceState::new(&device));
-
-        func(&mut client, &user, &mut seq);
-
-        assert!(tunnel_tunnel_ids.assigned.contains(&102));
-
-        process_user_event(
-            &client,
-            &pubkey,
-            &mut devices,
-            &mut user_tunnel_ips,
-            &mut tunnel_tunnel_ids,
-            &user,
-            &mut state_transitions,
-        );
-
-        assert!(!tunnel_tunnel_ids.assigned.contains(&102));
-
-        assert_eq!(state_transitions.len(), 1);
-        assert_eq!(state_transitions[state_transition], 1);
-    }
-
-    #[test]
-    fn test_process_user_event_deleting() {
-        do_test_process_user_event_deleting_or_pending_ban(
-            UserStatus::Deleting,
-            |user_service, user, seq| {
-                user_service
-                    .expect_execute_transaction()
-                    .times(1)
-                    .in_sequence(seq)
-                    .with(
-                        predicate::eq(DoubleZeroInstruction::DeactivateUser(UserDeactivateArgs {
-                            index: user.index,
-                            bump_seed: user.bump_seed,
-                        })),
-                        predicate::always(),
-                    )
-                    .returning(|_, _| Ok(Signature::new_unique()));
-            },
-            "user-deleting-to-deactivated",
-        );
-    }
-
-    #[test]
-    fn test_process_user_event_pending_ban() {
-        do_test_process_user_event_deleting_or_pending_ban(
-            UserStatus::PendingBan,
-            |user_service, user, seq| {
-                user_service
-                    .expect_execute_transaction()
-                    .times(1)
-                    .in_sequence(seq)
-                    .with(
-                        predicate::eq(DoubleZeroInstruction::BanUser(UserBanArgs {
-                            index: user.index,
-                            bump_seed: user.bump_seed,
-                        })),
-                        predicate::always(),
-                    )
-                    .returning(|_, _| Ok(Signature::new_unique()));
-            },
-            "user-pending-ban-to-banned",
-        );
-    }
-
-    #[test]
-    fn test_process_location_event() {
-        let mut locations = HashMap::new();
-        let pubkey = Pubkey::new_unique();
-        let location = Location {
-            account_type: AccountType::Location,
-            owner: Pubkey::new_unique(),
-            index: 0,
-            bump_seed: 42,
-            lat: 50.0,
-            lng: 20.0,
-            loc_id: 1234,
-            status: LocationStatus::Activated,
-            code: "nyc".to_string(),
-            name: "New York".to_string(),
-            country: "USA".to_string(),
-        };
-
-        process_location_event(&pubkey, &mut locations, &location);
-
-        assert!(locations.contains_key(&pubkey));
-        assert_eq!(*locations.get(&pubkey).unwrap(), location);
-    }
-
-    #[test]
-    fn test_process_exchange_event() {
-        let mut exchanges = HashMap::new();
-        let pubkey = Pubkey::new_unique();
-        let exchange = Exchange {
-            account_type: AccountType::Exchange,
-            owner: Pubkey::new_unique(),
-            index: 0,
-            bump_seed: 42,
-            lat: 50.0,
-            lng: 20.0,
-            loc_id: 1234,
-            status: ExchangeStatus::Activated,
-            code: "TestExchange".to_string(),
-            name: "TestName".to_string(),
-        };
-
-        process_exchange_event(&pubkey, &mut exchanges, &exchange);
-
-        assert!(exchanges.contains_key(&pubkey));
-        assert_eq!(*exchanges.get(&pubkey).unwrap(), exchange);
     }
 }
