+++ client/doublezero/Cargo.toml
@@ -34,3 +34,7 @@ hyperlocal = { version="0.8" }
 
 doublezero_sdk = { path = "../../smartcontract/sdk/rs", version = "0.1.0" }
 doublezero_cli = { path = "../../smartcontract/cli", version = "0.1.0" }
+assert_cmd = "2.0.17"
+
+[dev-dependencies]
+assert_cmd = "2.0.17"
