+++ smartcontract/cli/src/globalconfig/set.rs
@@ -15,6 +15,8 @@ pub struct SetGlobalConfigCliCommand {
     tunnel_tunnel_block: String,
     #[arg(long)]
     device_tunnel_block: String,
+    #[arg(long)]
+    multicasogroup_block: String,
 }
 
 impl SetGlobalConfigCliCommand {
@@ -27,6 +29,7 @@ impl SetGlobalConfigCliCommand {
             remote_asn: self.remote_asn,
             tunnel_tunnel_block: networkv4_parse(&self.tunnel_tunnel_block),
             user_tunnel_block: networkv4_parse(&self.device_tunnel_block),
+            multicastgroup_block: networkv4_parse(&self.multicasogroup_block),
         })?;
         writeln!(out, "Signature: {}", signature)?;
 
@@ -65,6 +68,7 @@ mod tests {
                 remote_asn: 5678,
                 tunnel_tunnel_block: ([10, 10, 0, 0], 16),
                 user_tunnel_block: ([10, 20, 0, 0], 16),
+                multicastgroup_block: ([224, 2, 0, 0], 4),
             }))
             .returning(move |_| Ok(signature));
 
@@ -75,6 +79,7 @@ mod tests {
             remote_asn: 5678,
             tunnel_tunnel_block: "10.10.0.0/16".to_string(),
             device_tunnel_block: "10.20.0.0/16".to_string(),
+            multicasogroup_block: "224.2.0.0/4".to_string(),
         }
         .execute(&client, &mut output);
         assert!(res.is_ok());
