+++ smartcontract/programs/dz-sla-program/src/processors/device/test.rs
@@ -53,6 +53,7 @@ mod device_test {
                 remote_asn: 65001,
                 tunnel_tunnel_block: ([10, 0, 0, 0], 24),
                 user_tunnel_block: ([10, 0, 0, 0], 24),
+                multicastgroup_block: ([224, 0, 0, 0], 4),
             }),
             vec![
                 AccountMeta::new(config_pubkey, false),
