+++ smartcontract/programs/dz-sla-program/src/processors/tunnel/test.rs
@@ -54,6 +54,7 @@ mod tunnel_test {
                 remote_asn: 65001,
                 tunnel_tunnel_block: ([10, 0, 0, 0], 24),
                 user_tunnel_block: ([10, 0, 0, 0], 24),
+                multicastgroup_block: ([10, 0, 0, 0], 24),
             }),
             vec![
                 AccountMeta::new(config_pubkey, false),
