+++ smartcontract/programs/dz-sla-program/src/processors/user/tests.rs
@@ -57,6 +57,7 @@ mod user_test {
                 remote_asn: 65001,
                 tunnel_tunnel_block: ([10, 0, 0, 0], 24),
                 user_tunnel_block: ([10, 0, 0, 0], 24),
+                multicastgroup_block: ([224, 0, 0, 0], 4),
             }),
             vec![
                 AccountMeta::new(config_pubkey, false),
