+++ smartcontract/programs/dz-sla-program/src/state/globalconfig.rs
@@ -11,23 +11,25 @@ use crate::{
 
 #[derive(BorshSerialize, BorshDeserialize, Debug, PartialEq, Clone)]
 pub struct GlobalConfig {
-    pub account_type: AccountType,      // 1
-    pub owner: Pubkey,                  // 32
-    pub bump_seed: u8,                  // 1
-    pub local_asn: u32,                 // 4
-    pub remote_asn: u32,                // 4
-    pub tunnel_tunnel_block: NetworkV4, // 5
-    pub user_tunnel_block: NetworkV4,   // 5
+    pub account_type: AccountType,       // 1
+    pub owner: Pubkey,                   // 32
+    pub bump_seed: u8,                   // 1
+    pub local_asn: u32,                  // 4
+    pub remote_asn: u32,                 // 4
+    pub tunnel_tunnel_block: NetworkV4,  // 5
+    pub user_tunnel_block: NetworkV4,    // 5
+    pub multicastgroup_block: NetworkV4, // 5
 }
 
 impl fmt::Display for GlobalConfig {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         write!(
             f,
-            "account_type: {}, owner: {}, local_asn: {}, remote_asn: {}, tunnel_tunnel_block: {}, user_tunnel_block: {}",
+            "account_type: {}, owner: {}, local_asn: {}, remote_asn: {}, tunnel_tunnel_block: {}, user_tunnel_block: {}, multicastgroup_block: {}",
             self.account_type, self.owner, self.local_asn, self.remote_asn,
             networkv4_to_string(&self.tunnel_tunnel_block),
-            networkv4_to_string(&self.user_tunnel_block)
+            networkv4_to_string(&self.user_tunnel_block),
+            networkv4_to_string(&self.multicastgroup_block)
         )
     }
 }
@@ -44,13 +46,14 @@ impl From<&[u8]> for GlobalConfig {
             remote_asn: parser.read_u32(),
             tunnel_tunnel_block: parser.read_networkv4(),
             user_tunnel_block: parser.read_networkv4(),
+            multicastgroup_block: parser.read_networkv4(),
         }
     }
 }
 
 impl GlobalConfig {
     pub fn size(&self) -> usize {
-        1 + 32 + 1 + 4 + 4 + 5 + 5
+        1 + 32 + 1 + 4 + 4 + 5 + 5 + 5
     }
 }
 
@@ -68,6 +71,7 @@ mod tests {
             remote_asn: 456,
             tunnel_tunnel_block: ([10, 0, 0, 1], 24),
             user_tunnel_block: ([10, 0, 0, 2], 24),
+            multicastgroup_block: ([224, 0, 0, 0], 4),
         };
 
         let data = borsh::to_vec(&val).unwrap();
@@ -79,6 +83,7 @@ mod tests {
         assert_eq!(val.remote_asn, val2.remote_asn);
         assert_eq!(val.tunnel_tunnel_block, val2.tunnel_tunnel_block);
         assert_eq!(val.user_tunnel_block, val2.user_tunnel_block);
+        assert_eq!(val.multicastgroup_block, val2.multicastgroup_block);
         assert_eq!(data.len(), val.size(), "Invalid Size");
     }
 }
