+++ smartcontract/sdk/rs/src/commands/globalconfig/set.rs
@@ -12,6 +12,7 @@ pub struct SetGlobalConfigCommand {
     pub remote_asn: u32,
     pub tunnel_tunnel_block: NetworkV4,
     pub user_tunnel_block: NetworkV4,
+    pub multicastgroup_block: NetworkV4,
 }
 
 impl SetGlobalConfigCommand {
@@ -27,6 +28,7 @@ impl SetGlobalConfigCommand {
                 remote_asn: self.remote_asn,
                 tunnel_tunnel_block: self.tunnel_tunnel_block,
                 user_tunnel_block: self.user_tunnel_block,
+                multicastgroup_block: self.multicastgroup_block,
             }),
             vec![
                 AccountMeta::new(pda_pubkey, false),
